requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.5
numpy==1.21.6
packaging==24.0
pandas==1.3.5
pluggy==1.2.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
